Replace more int indices by size_t

+++ src/mapping/RadialGeoMultiscaleMapping.cpp
@@ -62,10 +62,10 @@ void RadialGeoMultiscaleMapping::computeMapping()
       _vertexIndicesSpread.clear();
       _vertexIndicesSpread.reserve(output()->vertices().size());
       for (size_t i = 0; i < outSize; i++) {
-        auto vertexCoord = output()->vertices()[i].rawCoords()[effectiveCoordinate];
-        int  index       = 0;
+        auto   vertexCoord = output()->vertices()[i].rawCoords()[effectiveCoordinate];
+        size_t index       = 0;
         while (vertexCoord > axisMidpoints(index)) {
-          PRECICE_ASSERT(index + 1 < static_cast<int>(inSize));
+          PRECICE_ASSERT(index + 1 < inSize);
           ++index;
         }
         _vertexIndicesSpread.push_back(index);
@@ -102,10 +102,10 @@ void RadialGeoMultiscaleMapping::computeMapping()
       _vertexIndicesCollect.clear();
       _vertexIndicesCollect.reserve(input()->vertices().size());
       for (size_t i = 0; i < inSize; i++) {
-        auto vertexCoords = input()->vertices()[i].rawCoords()[effectiveCoordinate];
-        int  index        = 0;
+        auto   vertexCoords = input()->vertices()[i].rawCoords()[effectiveCoordinate];
+        size_t index        = 0;
         while (vertexCoords > axisMidpoints(index)) {
-          PRECICE_ASSERT(index + 1 < static_cast<int>(outSize));
+          PRECICE_ASSERT(index + 1 < outSize);
           ++index;
         }
         _vertexIndicesCollect.push_back(index);
